Settings: Fix turning off Wifi issue in Non-isRadioAllowed(RADIO_WIFI)
Under Non-isRadioAllowed(RADIO_WIFI) conditions, WifiEnabler
onSwitchToggled() cannot be turned on is in line with the
design expectations, but ignoring the situation of turning
off Wifi under Non-isRadioAllowed(RADIO_WIFI) and this may
exist.

In the above scenario, it will cause the inconsistency between
Wifi button state and the Wifi state.

Bug:253330621
Test:
atest SettingsRoboTests:com.android.settings.wifi.WifiEnablerTest
and Maunal

Signed-off-by: xuelin <xuelin@xiaomi.corp-partner.google.com>
Change-Id: I12e071eaad11e21b81d928d47d992cad7faf7ad2