Settings: Fix sometimes WiFi tethering multiple restarts if configuration changed.
When WiFi Hotspot is already started, if the user changes the configuration, the WiFi Hotspot will be restarted automatically. However, sometimes the restart will be triggered multiple times due to the tethering status change broadcast which is asynchronous and delayed.

Use the sync callback to replace the broadcast receiver.

Change-Id: Ib3b76fe1aa762c05959b5cda679aba51f7d94fe3
CRs-Fixed: 3570075